requirements.txt
haystack-ai==2.11.2
python-dotenv==1.0.1
sentence-transformers>=3.0.0